+++ main.tf
@@ -112,80 +112,125 @@ resource "aws_lb_target_group_attachment" "main" {
 resource "aws_lb_listener" "main" {
   count             = length(var.listeners) > 0 ? length(var.listeners) : 0
   load_balancer_arn = aws_lb.main.arn
-  port              = lookup(var.listeners[count.index], "port", null)
-  protocol          = lookup(var.listeners[count.index], "protocol", null)
-  alpn_policy       = lookup(var.listeners[count.index], "alpn_policy", null)
-  certificate_arn   = lookup(var.listeners[count.index], "protocol", null) == "HTTPS" || lookup(var.listeners[count.index], "protocol", null) == "TLS" && var.create_ssl_certificate ? join("", aws_acm_certificate.main.*.arn) : lookup(var.listeners[count.index], "certificate_arn", null)
-  ssl_policy        = lookup(var.listeners[count.index], "ssl_policy", null)
-  tags              = lookup(var.listeners[count.index], "tags", {})
+  port              = try(var.listeners[count.index]["port"], null)
+  protocol          = try(var.listeners[count.index]["protocol"], null)
+  alpn_policy       = try(var.listeners[count.index]["alpn_policy"], null)
+  certificate_arn   = (try(var.listeners[count.index]["protocol"], null) == "HTTPS" || try(var.listeners[count.index]["protocol"], null) == "TLS") && var.create_ssl_certificate ? join("", aws_acm_certificate.main.*.arn) : try(var.listeners[count.index]["certificate_arn"], null)
+  ssl_policy        = try(var.listeners[count.index]["ssl_policy"], null)
+  tags              = try(var.listeners[count.index]["tags"], {})
+
   dynamic "default_action" {
-    for_each = length(lookup(var.listeners[count.index], "default_action")) == 0 ? [] : [lookup(var.listeners[count.index], "default_action", {})]
+    for_each = try([var.listeners[count.index]["default_actions"]], [var.listeners[count.index]["default_action"]], [])
 
     content {
       type             = default_action.value.type
-      target_group_arn = contains([null, "", "forward"], lookup(default_action.value, "type", "")) ? aws_lb_target_group.main[lookup(default_action.value, "tg_index", count.index)].id : null
+      order            = try(default_action.value.order, null)
+      target_group_arn = try(default_action.value.target_group_arn, aws_lb_target_group.main[lookup(default_action.value, "tg_index")].id, null)
 
       dynamic "authenticate_oidc" {
-        for_each = length(lookup(default_action.value, "authenticate_oidc", {})) == 0 ? [] : [lookup(default_action.value, "authenticate_oidc", {})]
+        for_each = try([default_action.value.authenticate_oidc], [])
 
         content {
           # Max 10 extra params
-          authentication_request_extra_params = lookup(authenticate_oidc.value, "authentication_request_extra_params", null)
-          authorization_endpoint              = authenticate_oidc.value["authorization_endpoint"]
-          client_id                           = authenticate_oidc.value["client_id"]
-          client_secret                       = authenticate_oidc.value["client_secret"]
-          issuer                              = authenticate_oidc.value["issuer"]
-          token_endpoint                      = authenticate_oidc.value["token_endpoint"]
-          user_info_endpoint                  = authenticate_oidc.value["user_info_endpoint"]
-          on_unauthenticated_request          = lookup(authenticate_oidc.value, "on_unauthenticated_request", null)
-          scope                               = lookup(authenticate_oidc.value, "scope", null)
-          session_cookie_name                 = lookup(authenticate_oidc.value, "session_cookie_name", null)
-          session_timeout                     = lookup(authenticate_oidc.value, "session_timeout", null)
+          authentication_request_extra_params = try(authenticate_oidc.value.authentication_request_extra_params, null)
+          authorization_endpoint              = authenticate_oidc.value.authorization_endpoint
+          client_id                           = authenticate_oidc.value.client_id
+          client_secret                       = authenticate_oidc.value.client_secret
+          issuer                              = authenticate_oidc.value.issuer
+          token_endpoint                      = authenticate_oidc.value.token_endpoint
+          user_info_endpoint                  = authenticate_oidc.value.user_info_endpoint
+          on_unauthenticated_request          = try(authenticate_oidc.value.on_unauthenticated_request, null)
+          scope                               = try(authenticate_oidc.value.scope, null)
+          session_cookie_name                 = try(authenticate_oidc.value.session_cookie_name, null)
+          session_timeout                     = try(authenticate_oidc.value.session_timeout, null)
         }
       }
 
       dynamic "fixed_response" {
-        for_each = length(keys(lookup(default_action.value, "fixed_response", {}))) == 0 ? [] : [lookup(default_action.value, "fixed_response", {})]
+        for_each = try([default_action.value.fixed_response], [])
 
         content {
-          content_type = fixed_response.value["content_type"]
-          message_body = lookup(fixed_response.value, "message_body", null)
-          status_code  = lookup(fixed_response.value, "status_code", null)
+          content_type = fixed_response.value.content_type
+          message_body = try(fixed_response.value.message_body, null)
+          status_code  = try(fixed_response.value.status_code, null)
         }
       }
 
       dynamic "redirect" {
-        for_each = length(keys(lookup(default_action.value, "redirect", {}))) == 0 ? [] : [lookup(default_action.value, "redirect", {})]
+        for_each = try([default_action.value.redirect], [])
 
         content {
-          path        = lookup(redirect.value, "path", null)
-          host        = lookup(redirect.value, "host", null)
-          port        = lookup(redirect.value, "port", null)
-          protocol    = lookup(redirect.value, "protocol", null)
-          query       = lookup(redirect.value, "query", null)
-          status_code = redirect.value["status_code"]
+          path        = try(redirect.value.path, null)
+          host        = try(redirect.value.host, null)
+          port        = try(redirect.value.port, null)
+          protocol    = try(redirect.value.protocol, null)
+          query       = try(redirect.value.query, null)
+          status_code = redirect.value.status_code
         }
       }
 
-      ## For HTTPS listeners only
       dynamic "authenticate_cognito" {
-        for_each = length(keys(lookup(default_action.value, "authenticate_cognito", {}))) == 0 ? [] : [lookup(default_action.value, "authenticate_cognito", {})]
+        for_each = try([default_action.value.authenticate_cognito], [])
 
         content {
           # Max 10 extra params
-          authentication_request_extra_params = lookup(authenticate_cognito.value, "authentication_request_extra_params", null)
-          on_unauthenticated_request          = lookup(authenticate_cognito.value, "on_authenticated_request", null)
-          scope                               = lookup(authenticate_cognito.value, "scope", null)
-          session_cookie_name                 = lookup(authenticate_cognito.value, "session_cookie_name", null)
-          session_timeout                     = lookup(authenticate_cognito.value, "session_timeout", null)
-          user_pool_arn                       = authenticate_cognito.value["user_pool_arn"]
-          user_pool_client_id                 = authenticate_cognito.value["user_pool_client_id"]
-          user_pool_domain                    = authenticate_cognito.value["user_pool_domain"]
+          authentication_request_extra_params = try(authenticate_cognito.value.authentication_request_extra_params, null)
+          on_unauthenticated_request          = try(authenticate_cognito.value.on_authenticated_request, null)
+          scope                               = try(authenticate_cognito.value.scope, null)
+          session_cookie_name                 = try(authenticate_cognito.value.session_cookie_name, null)
+          session_timeout                     = try(authenticate_cognito.value.session_timeout, null)
+          user_pool_arn                       = authenticate_cognito.value.user_pool_arn
+          user_pool_client_id                 = authenticate_cognito.value.user_pool_client_id
+          user_pool_domain                    = authenticate_cognito.value.user_pool_domain
         }
       }
     }
   }
 }
+
+## Commented for other release so as not to lose the code
+# resource "aws_lb_listener_rule" "main" {
+#   for_each = {
+#     for idx, listener in var.listeners : idx => listener.rules
+#     if length(lookup(listener, "rules", [])) > 0
+#   }
+#
+#   listener_arn = try(each.value.listener_arn, aws_lb_listener.main[each.key].arn)
+#   priority     = try(each.value.priority, null)
+#
+#   dynamic "action" {
+#     for_each = [ for action in each.value.actions : action[0] ]
+#
+#     content {
+#       type  = action.value.type
+#       order = try(action.value.order, null)
+#
+#       dynamic "fixed_response" {
+#         for_each = try([action.value.fixed_response], [])
+#
+#         content {
+#           content_type = fixed_response.value.content_type
+#           message_body = try(fixed_response.value.message_body, null)
+#           status_code  = try(fixed_response.value.status_code, null)
+#         }
+#       }
+#     }
+#   }
+#
+#   dynamic "condition" {
+#     for_each = [ for condition in each.value.conditions : condition[0] ]
+#
+#     content {
+#       dynamic "host_header" {
+#         for_each = try([condition.value.host_header], [])
+#
+#         content {
+#           values = host_header.value.values
+#         }
+#       }
+#     }
+#   }
+# }
 
 #### Self Signed Certificate
 resource "tls_private_key" "main" {
@@ -220,11 +265,10 @@ resource "aws_acm_certificate" "main" {
   }
 }
 
-
 # Security group
 resource "aws_security_group" "main" {
   count       = var.load_balancer_type == "gateway" ? 0 : 1
-  name        = "${var.name}-security-group"
+  name        = try("${var.name}-security-group", "lb-sg")
   vpc_id      = var.vpc_id
   description = "Load balancer security group"
   tags = merge(